+++ test/utils/IUsdnProtocolHandler.sol
@@ -332,7 +332,7 @@ interface IUsdnProtocolHandler is IUsdnProtocol {
         address validator,
         PositionId memory posId,
         uint128 amountToClose,
-        uint128 userMinPrice,
+        uint256 userMinPrice,
         bytes calldata currentPriceData
     ) external returns (ClosePositionData memory data_, bool liquidated_);
